% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigQueryAccess.R
\name{getServiceAccountKey}
\alias{getServiceAccountKey}
\title{Function that returns the key for the user's pet service account}
\usage{
getServiceAccountKey(authEnv = "prod")
}
\arguments{
\item{authEnv}{the environment to auth against}
}
\value{
the user's pet service account key
}
\description{
Function that returns the key for the user's pet service account
}
