% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigQueryAccess.R
\name{getToken}
\alias{getToken}
\title{Function to get the token for bigquery access}
\usage{
getToken(authEnv = "prod")
}
\arguments{
\item{authEnv}{the environment to auth against}
}
\value{
the token to access BigQuery
}
\description{
Function to get the token for bigquery access
}
